+++ HISTORY.md
@@ -11,6 +11,7 @@
 
 ### 2021-02-06
 * lists, objects, threads: new RESHAPE primitive
+* lists: added internal naive (recursive)  version of CROSSPRODUCT
 
 ### 2021-02-06
 * simplified private list.range() method